Split trades methods into seperate converter file
