+++ freqtrade/data/converter/__init__.py
@@ -1,8 +1,28 @@
-# flake8: noqa: F401
 from freqtrade.data.converter.converter import (clean_ohlcv_dataframe, convert_ohlcv_format,
-                                                convert_trades_format, convert_trades_to_ohlcv,
                                                 ohlcv_fill_up_missing_data, ohlcv_to_dataframe,
                                                 order_book_to_dataframe, reduce_dataframe_footprint,
-                                                trades_convert_types, trades_df_remove_duplicates,
-                                                trades_dict_to_list, trades_list_to_df,
-                                                trades_to_ohlcv, trim_dataframe, trim_dataframes)
+                                                trim_dataframe, trim_dataframes)
+from freqtrade.data.converter.trade_converter import (convert_trades_format,
+                                                      convert_trades_to_ohlcv, trades_convert_types,
+                                                      trades_df_remove_duplicates,
+                                                      trades_dict_to_list, trades_list_to_df,
+                                                      trades_to_ohlcv)
+
+
+__all__ = [
+    'clean_ohlcv_dataframe',
+    'convert_ohlcv_format',
+    'ohlcv_fill_up_missing_data',
+    'ohlcv_to_dataframe',
+    'order_book_to_dataframe',
+    'reduce_dataframe_footprint',
+    'trim_dataframe',
+    'trim_dataframes',
+    'convert_trades_format',
+    'convert_trades_to_ohlcv',
+    'trades_convert_types',
+    'trades_df_remove_duplicates',
+    'trades_dict_to_list',
+    'trades_list_to_df',
+    'trades_to_ohlcv',
+]
